skills/user/research-data-hub/requirements.txt
# US Corporate Analytics - 依存関係

# コアライブラリ
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# データ取得
requests>=2.31.0
urllib3>=2.0.0

# 統計分析
statsmodels>=0.14.0
scikit-learn>=1.3.0

# 可視化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# ファイル処理
openpyxl>=3.1.0
xlsxwriter>=3.1.0
PyYAML>=6.0.0

# レポート生成
reportlab>=4.0.0
Pillow>=10.0.0

# テキスト処理
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ユーティリティ
python-dateutil>=2.8.0
pytz>=2023.3

# 品質保証
pytest>=7.4.0
pytest-cov>=4.1.0

# オプション（学術研究向け）
# linearmodels>=5.3  # パネルデータ分析
# arch>=6.2.0        # 時系列分析（ARCH/GARCHモデル）
